Improve performance of writing JSON values
The old code was exceedingly simple, as it went char-by-char and
decided whether to write it to the output stream as-is, or escaped.
This caused a _lot_ of stream writes of individual characters.

The new code instead looks for characters that need escaping, and
bulk-writes the non-escaped characters in between them. This leads
to 3x improvement for short value strings, and 10x+ improvement
for looong value strings.

In practice, we should expect the former rather than the latter,
but this is still nice improvement.